refactor : 상수 관리포인트 변경
